(SERVER-2215) Update cljs tools and use figwheel-sidecar
cljs-dashboard-widgets used older clojure, clojurescript, and cljsjs/d3 dependencies. I updated these to modern versions and added clj-parent to manage the rest of the dependencies in the project. Upgrading to d3 v4 broke some of the metrix-box animation, so I fixed that as well. Figwheel required a lein dependency to run lein figwheel, but upgrading to figwheel-sidecar allowed me to remove that dependency.